+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Change double and single signature creation to be a method on `NoteSecretKey` [#81]
+
 ## [0.15.0] - 2023-11-1
 
 ### Added
@@ -72,7 +76,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Add `CheckBytes` impl for `rkyv`ed structs
-- Add `rkyv` implementations behind feature [#69](https://github.com/dusk-network/schnorr/issues/69)
+- Add `rkyv` implementations behind feature [#69]
 
 ### Changed
 
@@ -193,8 +197,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fix byte truncation for BLS -> JubJub conversion
 
 <!-- ISSUES -->
+[#81]: https://github.com/dusk-network/schnorr/issues/81
 [#80]: https://github.com/dusk-network/schnorr/issues/80
 [#89]: https://github.com/dusk-network/schnorr/issues/89
+[#69]: https://github.com/dusk-network/schnorr/issues/69
 [#59]: https://github.com/dusk-network/schnorr/issues/59
 [#52]: https://github.com/dusk-network/schnorr/issues/52
 [#52]: https://github.com/dusk-network/schnorr/issues/52